pyproject.toml
[project]
name = "typst_importer"
version = "0.0.1"  # Placeholder; the real version is defined in blender_manifest.toml
description = "Blender extension to render Typst files"
license = { text = "AGPL-3.0-or-later" }
readme = "README.md"
dependencies = ["typst", "svg.path"]
requires-python = ">=3.11.0,<3.12"
keywords = ["blender", "python", "typst"]
maintainers = [
    { name = "Jan-Hendrik Müller", email = "jan-hendrik.mueller@uni-goettingen.de" }
]

[project.urls]
Homepage = "https://extensions.blender.org/add-ons/typst-importer/"
Repository = "https://github.com/kolibril13/blender_typst_importer"
# Documentation = "https://kolibril13.github.io/blender_typst_importer"

[project.optional-dependencies]
bpy = ["bpy>=4.2"]
test = ["pytest", "pytest-cov"]
dev = ["fake-bpy-module", "tomlkit"]
docs = ["jupyter"]

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[tool.setuptools]
packages = ["typst_importer"]
